python/requirements.txt
Faker==2.0.0
python-dateutil==2.8.0
six==1.12.0
text-unidecode==1.2
